Milestone3/requirements.txt
openai
pinecone-client
sentence-transformers
pandas
streamlit
requests
